+++ app/src/App.tsx
@@ -1,7 +1,5 @@
-import React, { useEffect, useState } from "react";
-import { ThemeProvider } from "@emotion/react";
+import { useEffect, useState } from "react";
 import styled from "@emotion/styled";
-import { createTheme, CssBaseline } from "@mui/material";
 
 import { DialogConfig } from "./components/Dialog/DialogConfig";
 import { DialogNoAPI } from "./components/Dialog/DialogNoAPI";
@@ -9,6 +7,7 @@ import { DialogToken } from "./components/Dialog/DialogToken";
 import { Footer } from "./components/Footer";
 import { ProcessingList } from "./components/Processing/ProcessingList";
 import { Results } from "./components/Results";
+import { useAuth } from "./provider/AuthProvider";
 import { ConfigProvider } from "./provider/ConfigProvider";
 import {
   ProcessingProvider,
@@ -28,45 +27,35 @@ declare module "@mui/material/styles/createPalette" {
   }
 }
 
-const darkTheme = createTheme({
-  palette: {
-    mode: "dark",
-    gold: "#a57c00",
-    alert: "#e47964",
-  },
-});
-
 function App() {
   const { apiError } = useProcessingProvider();
+  const { isAuthActive } = useAuth();
   const [appLoaded, setAppLoaded] = useState(false);
 
   useEffect(() => apiError && console.log(apiError), [apiError]);
-  useEffect(() => setAppLoaded(true), []);
+  useEffect(() => setAppLoaded(isAuthActive !== undefined), [isAuthActive]);
 
   if (!appLoaded) return null;
 
   return (
-    <ThemeProvider theme={darkTheme}>
-      <CssBaseline />
-      <ConfigProvider>
-        <main className="flex min-h-screen flex-col items-center justify-between">
-          <div className="relative">
-            <SearchProvider>
-              <ProcessingProvider>
-                <Content>
-                  <Results />
-                </Content>
-                <ProcessingList />
-                <DialogToken />
-                <DialogNoAPI />
-                <DialogConfig />
-              </ProcessingProvider>
-            </SearchProvider>
-          </div>
-        </main>
-        <Footer />
-      </ConfigProvider>
-    </ThemeProvider>
+    <ConfigProvider>
+      <main className="flex min-h-screen flex-col items-center justify-between">
+        <div className="relative">
+          <SearchProvider>
+            <ProcessingProvider>
+              <Content>
+                <Results />
+              </Content>
+              <ProcessingList />
+              <DialogToken />
+              <DialogNoAPI />
+              <DialogConfig />
+            </ProcessingProvider>
+          </SearchProvider>
+        </div>
+      </main>
+      <Footer />
+    </ConfigProvider>
   );
 }