+++ docker/run-dev.sh
@@ -1,4 +1,8 @@
+# AUTH JWT RANDOM SECRET
+export JWT_SECRET=`tr -dc A-Za-z0-9 </dev/urandom | head -c 15; echo`
+
 bash /home/app/build/docker/env.sh
+
 sleep 2
 yarn install
 yarn --cwd ./app install